fix: change from controller to use tooltop
