+++ packages/components/tooltip/src/index.ts
@@ -1,3 +1,2 @@
-export { default } from './TooltipController';
-export { default as Tooltip } from './TooltipController';
+export { default as Tooltip } from './Tooltip';
 export * from './types';
