+++ packages/components/tooltip/src/usecase/index.ts
@@ -1,0 +1,2 @@
+export { default as useHandlePosition } from './use-handle-position';
+export { default as useSelectAnchors } from './use-select-anchors';
